backend/foodgram_backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==3.2.16
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
dotenv==0.9.9
idna==3.10
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
gunicorn==20.1.0 
django-cors-headers==3.13.0
psycopg2-binary==2.9.3 
drf-extra-fields==3.7.0